requirements.txt
pillow~=10.4.0
numpy~=2.0.0